TestIsTest: move ill-formed tests into a testdata subpackage
This avoids triggering the go1.24 vet check for ill-formed
Test (etc) function names.

Fixes golang/go#68999

Reviewed-on: https://go-review.googlesource.com/c/playground/+/607835
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
